Struct matrix_sdk_sql::StateStore

pub struct StateStore<DB: SupportedDatabase> { /* private fields */ }

SQL State Storage for matrix-sdk

Implementations

impl<DB: SupportedDatabase> StateStore<DB>

pub async fn new(db: &Arc<Pool<DB>>) -> Result<Self> where
<DB as Database>::Connection: Migrate,

Create a new State Store and automtaically performs migrations

Errors

This function will return an error if the migration cannot be applied

pub async fn unlock(&mut self) -> Result<()> where
for<'a> <DB as HasArguments<'a>>::Arguments: IntoArguments<'a, DB>,
for<'c> &'c mut <DB as Database>::Connection: Executor<'c, Database = DB>,
for<'c, 'a> &'a mut Transaction<'c, DB>: Executor<'a, Database = DB>,
for<'a> &'a [u8]: BorrowedSqlType<'a, DB>,
for<'a> &'a str: BorrowedSqlType<'a, DB>,
Vec<u8>: SqlType<DB>,
String: SqlType<DB>,
bool: SqlType<DB>,
Vec<u8>: SqlType<DB>,
Option<String>: SqlType<DB>,
Json<Raw<AnyGlobalAccountDataEvent>>: SqlType<DB>,
Json<Raw<PresenceEvent>>: SqlType<DB>,
Json<SyncRoomMemberEvent>: SqlType<DB>,
Json<MinimalRoomMemberEvent>: SqlType<DB>,
Json<Raw<AnySyncStateEvent>>: SqlType<DB>,
Json<Raw<AnyRoomAccountDataEvent>>: SqlType<DB>,
Json<RoomInfo>: SqlType<DB>,
Json<Receipt>: SqlType<DB>,
Json<Raw<AnyStrippedStateEvent>>: SqlType<DB>,
Json<StrippedRoomMemberEvent>: SqlType<DB>,
Json<MemberEvent>: SqlType<DB>,
for<'a> &'a str: ColumnIndex<<DB as Database>::Row>,

Unlocks the e2e encryption database

Errors

This function will fail if the database could not be unlocked

pub async fn unlock_with_passphrase(&mut self, passphrase: &str) -> Result<()> where
for<'a> <DB as HasArguments<'a>>::Arguments: IntoArguments<'a, DB>,
for<'c> &'c mut <DB as Database>::Connection: Executor<'c, Database = DB>,
for<'c, 'a> &'a mut Transaction<'c, DB>: Executor<'a, Database = DB>,
for<'a> &'a [u8]: BorrowedSqlType<'a, DB>,
for<'a> &'a str: BorrowedSqlType<'a, DB>,
Vec<u8>: SqlType<DB>,
String: SqlType<DB>,
bool: SqlType<DB>,
Vec<u8>: SqlType<DB>,
Option<String>: SqlType<DB>,
Json<Raw<AnyGlobalAccountDataEvent>>: SqlType<DB>,
Json<Raw<PresenceEvent>>: SqlType<DB>,
Json<SyncRoomMemberEvent>: SqlType<DB>,
Json<MinimalRoomMemberEvent>: SqlType<DB>,
Json<Raw<AnySyncStateEvent>>: SqlType<DB>,
Json<Raw<AnyRoomAccountDataEvent>>: SqlType<DB>,
Json<RoomInfo>: SqlType<DB>,
Json<Receipt>: SqlType<DB>,
Json<Raw<AnyStrippedStateEvent>>: SqlType<DB>,
Json<StrippedRoomMemberEvent>: SqlType<DB>,
Json<MemberEvent>: SqlType<DB>,
for<'a> &'a str: ColumnIndex<<DB as Database>::Row>,

Unlocks the e2e encryption database with password

Errors

This function will fail if the passphrase is wrong